+++ extensions/system-monitor/src/menubar-system-monitor.tsx
@@ -16,6 +16,8 @@ export default function Command() {
   const { customIconUrl } = getPreferenceValues<Preferences.MenubarSystemMonitor>();
   const { displayModeCpu, displayModeBattery, displayModeDisk, displayModeMemory } =
     getPreferenceValues<ExtensionPreferences>();
+  const { cpuMenubarFormat, memoryMenubarFormat, powerMenubarFormat, networkMenubarFormat, diskMenubarFormat } =
+    getPreferenceValues<Preferences.MenubarSystemMonitor>();
 
   const {
     data: systemInfo,
@@ -82,6 +84,21 @@ export default function Command() {
     };
   });
 
+  const formatTags = (
+    formatString: string,
+    value: string = "",
+    total: string = "",
+    percent: string = "",
+    displayMode: string = "free",
+  ): string => {
+    return formatString
+      .replaceAll("<BR>", `\n`)
+      .replaceAll("<MODE>", displayMode === "free" ? "Free" : "Used")
+      .replace("<VALUE>", value)
+      .replace("<TOTAL>", total)
+      .replace("<PERCENT>", percent);
+  };
+
   const { data: batteryData, revalidate: revalidateBattery } = usePromise(async () => {
     const batteryData = await getBatteryData();
     const isOnAC = !batteryData.isCharging && batteryData.fullyCharged;
@@ -125,9 +142,15 @@ export default function Command() {
             key={index}
             title={disk.diskName}
             subtitle={
-              displayModeDisk === "free"
-                ? `${disk.totalAvailableStorage} GB available of ${disk.totalSize} GB`
-                : `${disk.usedStorage} GB used of ${disk.totalSize} GB`
+              disk
+                ? formatTags(
+                    diskMenubarFormat,
+                    displayModeDisk === "free" ? disk.totalAvailableStorage : disk.usedStorage,
+                    disk.totalSize,
+                    "",
+                    displayModeDisk,
+                  )
+                : "Loading…"
             }
             icon={Icon.HardDrive}
             onAction={() => runAppleScript(openActivityMonitorAppleScript(4))}
@@ -138,7 +161,17 @@ export default function Command() {
       <MenuBarExtra.Section title="CPU">
         <MenuBarExtra.Item
           title="CPU Usage"
-          subtitle={cpuUsage ? `${displayModeCpu === "free" ? 100 - +cpuUsage : cpuUsage} %` : "Loading..."}
+          subtitle={
+            cpuUsage
+              ? formatTags(
+                  cpuMenubarFormat,
+                  "",
+                  "",
+                  `${displayModeCpu === "free" ? 100 - +cpuUsage : cpuUsage}`,
+                  displayModeCpu,
+                )
+              : "Loading..."
+          }
           icon={Icon.Monitor}
           onAction={() => runAppleScript(openActivityMonitorAppleScript(1))}
         />
@@ -148,11 +181,21 @@ export default function Command() {
         <MenuBarExtra.Item
           title="Memory Usage"
           subtitle={
-            !memoryUsage
-              ? "Loading..."
-              : displayModeMemory === "free"
-                ? `${memoryUsage?.freeMemPercentage} % (~ ${memoryUsage?.freeMem} GB)`
-                : `${100 - +memoryUsage.freeMemPercentage} % (~ ${+memoryUsage.totalMem - +memoryUsage.freeMem} GB)`
+            memoryUsage
+              ? displayModeMemory === "free"
+                ? formatTags(
+                    memoryMenubarFormat,
+                    memoryUsage.freeMem,
+                    memoryUsage.totalMem,
+                    memoryUsage.freeMemPercentage,
+                  )
+                : formatTags(
+                    memoryMenubarFormat,
+                    (+memoryUsage.totalMem - +memoryUsage.freeMem).toString(),
+                    memoryUsage.totalMem,
+                    (100 - +memoryUsage.freeMemPercentage).toString(),
+                  )
+              : "Loading…"
           }
           icon={Icon.MemoryChip}
           onAction={() => runAppleScript(openActivityMonitorAppleScript(2))}
@@ -162,9 +205,13 @@ export default function Command() {
       <MenuBarExtra.Section title="Network">
         <MenuBarExtra.Item
           title="Network Usage"
-          subtitle={`↓ ${networkUsage?.download !== undefined ? formatBytes(networkUsage.download) : "0 B"}/s ↑ ${
-            networkUsage?.upload !== undefined ? formatBytes(networkUsage.upload) : "0 B"
-          }/s`}
+          subtitle={
+            networkUsage
+              ? formatTags(networkMenubarFormat)
+                  .replace("<UP>", formatBytes(networkUsage.upload))
+                  .replace("<DOWN>", formatBytes(networkUsage.download))
+              : "Loading…"
+          }
           icon={Icon.Network}
           onAction={() => runAppleScript(openActivityMonitorAppleScript(5))}
         />
@@ -174,9 +221,16 @@ export default function Command() {
         <MenuBarExtra.Item
           title="Battery"
           subtitle={
-            !batteryData?.batteryData.batteryLevel
-              ? "Loading..."
-              : `${displayModeBattery === "free" ? batteryData.batteryData.batteryLevel : 100 - +batteryData.batteryData.batteryLevel} %`
+            batteryData
+              ? formatTags(
+                  powerMenubarFormat,
+                  "",
+                  "",
+                  displayModeBattery === "free"
+                    ? batteryData.batteryData.batteryLevel
+                    : (100 - +batteryData.batteryData.batteryLevel).toString(),
+                )
+              : "Loading…"
           }
           icon={Icon.Plug}
           onAction={() => runAppleScript(openActivityMonitorAppleScript(3))}