+++ extensions/system-monitor/CHANGELOG.md
@@ -1,7 +1,13 @@
 # System Monitor Changelog
 
-## [Improvements] - {PR_MERGE_DATE}
-- updated free and used preference to be per-module for cpu, memory, disk and battery usage
+## [New Additions & Chore] - {PR_MERGE_DATE}
+- Added customisable tags for menubar entries
+    - Universal tags
+        - `<BR>` for line breaks
+        - `<MODE>` for display mode(toggles between "Free" and "Used")
+    - Module specific tags can be seen by hovering over the preferences text box
+- Made Loading tags use `…` consistently instead of `...`
+- Updated free and used preference to be per-module for cpu, memory, disk and battery usage
 
 ## [Toggle Display Mode + Modernize + Add README] - 2026-01-19